requirements.txt
fastapi==0.115.6
google-cloud-aiplatform==1.74.0
gunicorn==23.0.0
httpx==0.28.1
openai==1.59.6
pydantic-settings==2.7.0
uvicorn==0.32.1
